implementing parallax
Added a method to calculate delta values of transform.position for getting change in value between each frame.
parallax effect seems to be less. try multiplying by 2.

-Default method is using camera.transform to get position vector delta values btw frames.
and modify delta values by multiplying it by parallaxFactor and then adding that to transform.position of the background gameObject.

for gameObjects far away in background give parallaxFactor value closer to 1.0
(if u give 1 it will become constant on the camera view.)
for gameObject near in background which u want to pass by faster give parallaxFactor value closer to 0